# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Scrapor API
  version: 1.0.0
  description: |
    Scrapor multi-tenant video platform API.

    ## Authentication

    All documented endpoints accept an API key via:
    - Header: `Authorization: Bearer spk_ro_{id}_{secret}` or `spk_fa_{id}_{secret}`
    - Query: `?apiKey=spk_…`

    **Key types:**
    - `READ_ONLY` (`spk_ro_…`) — GET/HEAD/OPTIONS only, including full video stream via `GET /api/videos/{id}/stream`
    - `FULL_ACCESS` (`spk_fa_…`) — all HTTP methods on all routes (except `/api/api-keys/*` which is admin-session only)

    Browser sessions (cookies) also work for the admin panel and public site. When an API key is present, its type and method permissions are enforced.

    This spec and `/api/docs` require a valid API key (any type).
servers:
  - url: /
    description: Current host
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token with format `spk_ro_{keyNumber}_{secret}` or `spk_fa_{keyNumber}_{secret}`.
        Alternatively pass `apiKey` as a query parameter.
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/video/{id}:
    get:
      tags: [videos]
      summary: Stream local dev video file
      description: Serves video from local `downloads/` folder with range request support. READ_ONLY accessible.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Video stream }
        "206": { description: Partial content (range) }
        "404": { description: Not found }
...
  /api/creator/apply:
    post:
      tags: [creator]
      summary: Submit creator application
      description: Requires logged-in user session (not API key alone).
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [displayName]
              properties:
                displayName: { type: string }
                desiredSlug: { type: string }
                bio: { type: string }
      responses:
        "200": { description: Application submitted }
        "401": { description: Login required }
...
  /api/admin/videos/{id}:
    patch:
      tags: [admin]
      summary: Update video metadata
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                tags: { type: array, items: { type: string } }
                pornstars: { type: array, items: { type: string } }
                isDeleted: { type: boolean }
      responses:
        "200": { description: Updated }
    delete:
      tags: [admin]
      summary: Soft-delete video
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
  /api/admin/videos/{id}/playback:
    post:
      tags: [admin]
      summary: Admin preview stream
      description: Mints admin preview CDN URL (bypasses ads and soft-delete gate).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Preview URL }
...
  /api/chats:
    get:
      tags: [assistant]
      summary: Search chat threads
      parameters:
        - name: query
          in: query
          schema: { type: string }
      responses:
        "200": { description: Thread list }
    post:
      tags: [assistant]
      summary: Sync chat thread
      responses:
        "200": { description: Synced }
...
  /api/contexts/{id}:
    put:
      tags: [assistant]
      summary: Update context
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Updated }
...
  /api/agent-scripts/{id}:
    put:
      tags: [assistant]
      summary: Update script
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Updated }
...
  /api/api-keys:
    get:
      tags: [admin]
      security: []
      summary: List API keys (admin session only)
      description: Requires admin browser session. API keys cannot manage themselves.
      responses:
        "200": { description: Key list }
        "401": { description: Unauthorized }
    post:
      tags: [admin]
      security: []
      summary: Create API key (admin session only)
      description: Returns the raw key once in the response. Admin session required.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name: { type: string }
                type: { type: string, enum: [READ_ONLY, FULL_ACCESS] }
      responses:
        "200": { description: Key created with one-time plaintext }
  /api/api-keys/{id}:
    patch:
      tags: [admin]
      security: []
      summary: Rename API key
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Renamed }
    delete:
      tags: [admin]
      security: []
      summary: Revoke API key
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Revoked }
...